[fix] 記事の変換処理を共通処理に移行
